requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
facebook-sdk==3.1.0
get==2019.4.13
idna==2.9
Naked==0.1.31
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
pycrypto==2.6.1
pytz==2019.3
PyYAML==5.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
shellescape==3.8.1
sqlparse==0.3.1
urllib3==1.25.8
